git-svn-id: svn://svn.code.sf.net/p/irrlicht/code/trunk@1306 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/CReadFile.cpp
@@ -37,7 +37,7 @@ s32 CReadFile::read(void* buffer, u32 sizeToRead)
 	if (!isOpen())
 		return 0;
 
-	return fread(buffer, 1, sizeToRead, File);
+	return (s32)fread(buffer, 1, sizeToRead, File);
 }
